Create README.md
A Leetcode problem. In this question we have to find out the minimum number of floors to find the threshold floor of a building. Threshold floor means the floor from which we drop an egg, it won't break and the egg won't break if we throw the egg from the floors below that threshold floor but the egg will break if we try to throw the egg from the floors which are above the threshold floor.
Eg: 3 eggs and 5 floors
Output: Minimum number of ways are: 3